linted and samll changes #3968

+++ sechub-web-ui/test-setups/docker-setups/setup.sh
@@ -10,7 +10,13 @@ srcdir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )
 cd $srcdir 
 
 # Copy the .env file to the current directory
-cp ../../.env .
+if [ -e x.txt ]
+then
+    echo "Using existsing .env file"
+else
+    echo "Coping .env file from web-ui directory"
+    cp ../../.env .
+fi
 
 # Source the .env file to load environment variables
 set -a
